requirements.txt
# PyRAG - Engineering Standards Assistant
# Python 3.10+ required (tested with Python 3.13)

# Core RAG Framework (Python 3.13 compatible)
llama-index
llama-index-core
llama-index-embeddings-openai
llama-index-llms-openai
llama-index-llms-deepseek
llama-index-vector-stores-qdrant

# Vector Database
qdrant-client

# PDF Processing
PyMuPDF
pymupdf4llm

# API Framework
fastapi
uvicorn[standard]
pydantic
pydantic-settings

# Utilities
python-dotenv
openai
tiktoken
loguru
tqdm

# Graph Visualization
pyvis
plotly
pywebview  # Embedded HTML viewer for interactive graphs

# Modern Windows GUI
customtkinter
pillow

# PDF Export
reportlab

# Word Export
python-docx

# Graph Visualization
networkx
matplotlib

# Hybrid Search & Reranking
rank-bm25

# Optional: Advanced table processing
tabulate
pandas
